ECER-1407: PR feedbacks resolved
