+++ src/ECER.Clients.RegistryPortal/ECER.Clients.RegistryPortal.Server/ConfigurationMapper.cs
@@ -1,0 +1,11 @@
+﻿using AutoMapper;
+
+namespace ECER.Clients.RegistryPortal.Server;
+
+public class ConfigurationMapper : Profile
+{
+  public ConfigurationMapper()
+  {
+    CreateMap<Managers.Admin.Contract.MetadataResources.Province, Province>();
+  }
+}
