--- src/ECER.Clients.RegistryPortal/ECER.Clients.RegistryPortal.Server/MetadataResources/MetadataResourceMapper.cs
@@ -1,11 +1,0 @@
-﻿using AutoMapper;
-
-namespace ECER.Clients.RegistryPortal.Server.MetadataResources;
-
-public class MetadataResourceMapper : Profile
-{
-  public MetadataResourceMapper()
-  {
-    CreateMap<Managers.Registry.Contract.MetadataResources.Province, Province>();
-  }
-}
