+++ src/ECER.Managers.Admin.Contract/MetadataResources/Contract.cs
@@ -1,0 +1,18 @@
+﻿using MediatR;
+
+namespace ECER.Managers.Admin.Contract.MetadataResources;
+
+/// <summary>
+/// Invokes provinces use case
+/// </summary>
+public record ProvincesQuery : IRequest<ProvincesQueryResults>
+{
+  public string? ById { get; set; }
+}
+/// <summary>
+/// Container for <see cref="ProvincesQuery"/> results
+/// </summary>
+/// <param name="Items">The </param>
+public record ProvincesQueryResults(IEnumerable<Province> Items);
+
+public record Province(string ProvinceId, string ProvinceName);
